Add get_num_threads function
While it is possible to change the number of threads used by FFTW plans with
`set_num_threads`, it is impossible to see what the current value is. This is
made difficult because FFTW3 does not, as far as I know, provide a way to do
this. However, caching the last argument passed to `set_num_threads` as a
constant in FFTW.jl allows later access. I have added a constant,
`PLAN_NUM_THREADS` to keep track of the last argument to `set_num_threads`,
modified `set_num_threads` to update this constant, and added a function
`get_num_threads` to access the constant.