Require Julia 1.6+, use Preferences, eliminate mutable state
This adds support for `Preferences.jl` as the backing store for setting
a computation backend (currently only supports `fftw` and `mkl` as valid
choices).  The convenience method `set_provider!()` is given, but it is
little more than a wrapper around `Preferences.set_preferences!()`.
Users can alter their `LocalPreferences.toml` or `Project.toml` files by
hand as well; in the event that an illegal value is provided, an error
will be displayed and `fftw` will be chosen as the default value.

Preferences is a relatively new subsystem in Julia and it has some
lacking support in `Pkg`; so for now we are adding a separate
`test/Project.toml`, which will have preferences added into it on CI, as
there is a bug in propagating preferences when a single `Project.toml`
is used with extra dependencies for the tests.