Revert back to the standard form for ccall which supports LazyLibrary improvements
We had initially planned a breaking change design for LazyLibrary, which
this package was updated for, but later we realized a better design
which allows going back to the original version of this code. The
previous form is still very poorly supported in the compiler and was
still rather buggy internally.

Revert "Replace `__init__` with `OncePerProcess` on supported Julia versions (#316)"
This reverts commit dbcb5d210be95a0edac920646884784da87f7cab.

Revert "Make FFTW relocatable for PackageCompiler"
This reverts commit e96270a112ff2c16e19fc6aaac1ec429668ebe2f.